payment-service/requirements.txt
Flask==2.3.3
pika==1.3.2
requests==2.31.0
python-dotenv==1.0.0
pyjwt==2.8.0
pytest==7.4.0
Flask-Cors==4.0.0 